requirements.txt
aiogram>=3.4.0
google-generativeai>=0.8.0
motor>=3.4.0
pydantic-settings>=2.2.1
